requirements.txt
# Core dependencies
Django==5.0.2
django-rest-framework==0.1.0
python-chess==1.999
openai==1.12.0

# Development dependencies
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
